improvement of headers

+++ binding/base64.c
@@ -1,3 +1,20 @@
+/*
+ * Copyright (C) 2017 "IoT.bzh"
+ * Author: José Bollo <jose.bollo@iot.bzh>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include <stdlib.h>
 #include <stdint.h>
 #include <string.h>
@@ -39,7 +56,7 @@ char *base64_encode_array_variant(const char * const *args, size_t count, const 
 		s = args[n++];
 		c = s[i = 0];
 		while (c) {
-			x = (x << 8) | (uint16_t)(uint8_t)c;
+			x = (uint16_t)((uint16_t)(x << 8) | (uint16_t)(uint8_t)c);
 			switch (j & 3) {
 			case 0:
 				buffer[j++] = eb64((x >> 2) & 63, v);
